+++ remote/src/app/pages/app-remote/app-remote.tsx
@@ -17,7 +17,7 @@ import {
   DeckdeckgoEventSlide,
   DeckdeckgoEventNextPrevSlide,
   DeckdeckgoEventDeckReveal,
-  ConnectionState
+  ConnectionState,
 } from '@deckdeckgo/types';
 
 // Utils
@@ -31,7 +31,7 @@ import {ParseAttributesUtils} from '../../utils/parse-attributes.utils';
 
 @Component({
   tag: 'app-remote',
-  styleUrl: 'app-remote.scss'
+  styleUrl: 'app-remote.scss',
 })
 export class AppRemote {
   @Element() el: HTMLElement;
@@ -88,7 +88,7 @@ export class AppRemote {
       if (state === ConnectionState.CONNECTED) {
         this.communicationService.emit({
           type: DeckdeckgoEventType.SLIDES_REQUEST,
-          emitter: DeckdeckgoEventEmitter.APP
+          emitter: DeckdeckgoEventEmitter.APP,
         });
       }
     });
@@ -328,7 +328,7 @@ export class AppRemote {
     this.communicationService.emit({
       type: type,
       emitter: DeckdeckgoEventEmitter.APP,
-      slideAnimation: slideAnimation
+      slideAnimation: slideAnimation,
     });
   }
 
@@ -453,7 +453,7 @@ export class AppRemote {
     this.communicationService.emit({
       type: DeckdeckgoEventType.SLIDE_ACTION,
       emitter: DeckdeckgoEventEmitter.APP,
-      action: this.action
+      action: this.action,
     });
 
     await this.actionPlayPause();
@@ -489,7 +489,7 @@ export class AppRemote {
 
   private async openConnectModal() {
     const modal: HTMLIonModalElement = await modalController.create({
-      component: 'app-remote-connect'
+      component: 'app-remote-connect',
     });
 
     modal.onDidDismiss().then(async (detail: OverlayEventDetail) => {
@@ -507,8 +507,8 @@ export class AppRemote {
     const modal: HTMLIonModalElement = await modalController.create({
       component: 'app-remote-slide-picker',
       componentProps: {
-        slides: this.slides
-      }
+        slides: this.slides,
+      },
     });
 
     modal.onDidDismiss().then(async (detail: OverlayEventDetail) => {
@@ -526,7 +526,7 @@ export class AppRemote {
     this.communicationService.emit({
       type: DeckdeckgoEventType.SLIDE_TO,
       emitter: DeckdeckgoEventEmitter.APP,
-      index: newSlideIndex
+      index: newSlideIndex,
     });
   }
 
@@ -552,15 +552,15 @@ export class AppRemote {
           role: 'cancel',
           handler: () => {
             // Nothing
-          }
+          },
         },
         {
           text: 'Yes',
           handler: async () => {
             await this.disconnect();
-          }
-        }
-      ]
+          },
+        },
+      ],
     });
 
     await alert.present();
@@ -641,7 +641,7 @@ export class AppRemote {
             {this.renderExtraActions()}
           </div>
           <app-notes></app-notes>
-        </main>
+        </main>,
       ];
     } else if (this.connectionState !== ConnectionState.DISCONNECTED) {
       let text: string = 'Not connected';